Configured variables from int to float to make the last commit fully possible.

+++ Assets/Scripts/AI_Enemy/EnemyHealth.cs
@@ -6,7 +6,7 @@ using UnityEngine.AI;
 public class EnemyHealth : MonoBehaviour
 {
     public int startingHealth = 100;            
-    public int currentHealth;                   
+    public float currentHealth;                   
     public float sinkSpeed = 2.5f;              
     public int scoreValue = 10;                 
     public AudioClip deathClip;
@@ -37,7 +37,7 @@ public class EnemyHealth : MonoBehaviour
 
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         if (isDead)
             return;
